Converted project to .net Standard 1.6. Seperated Examples from library.
