+++ CoreRCON.sln
@@ -1,16 +1,15 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
+# Visual Studio 15
+VisualStudioVersion = 15.0.26430.16
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{600446C9-F807-4B19-807C-DDDE625CE558}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{A25CF464-1990-4F9D-92E4-E69BC9B9D137}"
-	ProjectSection(SolutionItems) = preProject
-		global.json = global.json
-	EndProjectSection
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CoreRCON", "src\CoreRCON\CoreRCON.csproj", "{CDFCA475-8F06-424B-9CF7-7D1E5D21F6D5}"
 EndProject
-Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "CoreRCON", "src\CoreRCON\CoreRCON.xproj", "{CDFCA475-8F06-424B-9CF7-7D1E5D21F6D5}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TF2Example", "src\TF2Example\TF2Example.csproj", "{4157A4E6-B76E-4BC5-BE32-2171EA3A0BE1}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MCExample", "src\MCExample\MCExample.csproj", "{A7392D55-A455-4415-AD20-9298C5CFE7CB}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -22,11 +21,21 @@ Global
 		{CDFCA475-8F06-424B-9CF7-7D1E5D21F6D5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CDFCA475-8F06-424B-9CF7-7D1E5D21F6D5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CDFCA475-8F06-424B-9CF7-7D1E5D21F6D5}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4157A4E6-B76E-4BC5-BE32-2171EA3A0BE1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4157A4E6-B76E-4BC5-BE32-2171EA3A0BE1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4157A4E6-B76E-4BC5-BE32-2171EA3A0BE1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4157A4E6-B76E-4BC5-BE32-2171EA3A0BE1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A7392D55-A455-4415-AD20-9298C5CFE7CB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A7392D55-A455-4415-AD20-9298C5CFE7CB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A7392D55-A455-4415-AD20-9298C5CFE7CB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A7392D55-A455-4415-AD20-9298C5CFE7CB}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{CDFCA475-8F06-424B-9CF7-7D1E5D21F6D5} = {600446C9-F807-4B19-807C-DDDE625CE558}
+		{4157A4E6-B76E-4BC5-BE32-2171EA3A0BE1} = {600446C9-F807-4B19-807C-DDDE625CE558}
+		{A7392D55-A455-4415-AD20-9298C5CFE7CB} = {600446C9-F807-4B19-807C-DDDE625CE558}
 	EndGlobalSection
 EndGlobal
