--- src/CoreRCON/project.json
@@ -1,48 +1,0 @@
-{
-	"buildOptions": {
-		"emitEntryPoint": true,
-		"optimize": true
-	},
-	"packOptions": {
-		"owners": [ "Kredit" ],
-		"repository": {
-			"type": "git",
-			"url": "https://github.com/ScottKaye/CoreRCON"
-		},
-		"requireLicenseAcceptance": false,
-		"summary": "A .NET Core implementation of Valve's RCON spec and SRCDS's logaddress function.",
-		"iconUrl": "https://cdn.rawgit.com/ScottKaye/CoreRCON/master/logo.png",
-		"tags": [ "valve", "rcon", "logaddress", "srcds" ]
-	},
-	"name": "CoreRCON",
-	"version": "2.0.0",
-	"title": "CoreRCON",
-	"authors": [ "Scott Kaye" ],
-	"projectUrl": "https://github.com/ScottKaye/CoreRCON",
-	"description": "A .NET Core implementation of Valve's RCON spec and SRCDS's logaddress function.",
-	"scripts": {
-		"postcompile": [
-			"dotnet pack --no-build --configuration %compile:Configuration%"
-		]
-	},
-	"frameworks": {
-		"netcoreapp1.0": {
-			"imports": "dnxcore50",
-			"dependencies": {
-				"Microsoft.NETCore.App": {
-					"type": "platform",
-					"version": "1.0.0"
-				}
-			}
-		},
-		"netcoreapp1.1": {
-			"imports": "dnxcore50",
-			"dependencies": {
-				"Microsoft.NETCore.App": {
-					"type": "platform",
-					"version": "1.1.0"
-				}
-			}
-		}
-	}
-}
