+++ src/TF2Example/TF2Example.csproj
@@ -1,0 +1,13 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp1.1</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\CoreRCON\CoreRCON.csproj" />
+    <ProjectReference Include="..\src\CoreRCON\CoreRCON.csproj" />
+  </ItemGroup>
+
+</Project>
